+++ CHANGELOG.md
@@ -6,13 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.1] - 2019-08-15
 ### Added
 - Add site_builder_console:responsive_image:create command
 
 ### Changed
 - Show success messages as success
 
-## [0.2.0]
+## [0.2.0] - 2019-07-26
 ### Added
 - Create `form` and `view` display configurations for new bundles
 - Add further options for node bundles
@@ -35,7 +37,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add site_builder_console:field:create command
 - Add site_builder_console:field:delete command
 
-[Unreleased]: https://github.com/wongjn/site_builder_console/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/wongjn/site_builder_console/compare/0.2.1...HEAD
+[0.2.1]: https://github.com/wongjn/site_builder_console/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/wongjn/site_builder_console/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/wongjn/site_builder_console/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/wongjn/site_builder_console/compare/0.1.0...0.1.1